Better input validation
Redefinition and undefined symbols / operators are no longer valid
input.
Preventing the use of undefined symbols prevents recursions in the
generated code. This could have drawbacks so maybe this will be
revisited in the future.

These changes are encapsulated by renaming the previous simple visitor
and some code changes to it.